thesis/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pymongo
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
uvicorn
mailjet_rest